+++ pkg/proxy/proxy.go
@@ -75,11 +75,10 @@ type Proxy struct {
 	password   string
 	cachedOnly bool
 	debrid     debrid.Service
-	cache      *common.Cache
 	logger     *log.Logger
 }
 
-func NewProxy(config common.Config, deb *debrid.DebridService, cache *common.Cache) *Proxy {
+func NewProxy(config common.Config, deb *debrid.DebridService) *Proxy {
 	cfg := config.Proxy
 	port := cmp.Or(os.Getenv("PORT"), cfg.Port, "8181")
 	return &Proxy{
@@ -90,7 +89,6 @@ func NewProxy(config common.Config, deb *debrid.DebridService, cache *common.Cac
 		password:   cfg.Password,
 		cachedOnly: *cfg.CachedOnly,
 		debrid:     deb.Get(),
-		cache:      cache,
 		logger:     common.NewLogger("Proxy", os.Stdout),
 	}
 }