+++ pkg/qbit/main.go
@@ -8,8 +8,8 @@ import (
 	"goBlack/pkg/qbit/server"
 )
 
-func Start(ctx context.Context, config *common.Config, deb *debrid.DebridService, cache *common.Cache) error {
-	srv := server.NewServer(config, deb, cache)
+func Start(ctx context.Context, config *common.Config, deb *debrid.DebridService) error {
+	srv := server.NewServer(config, deb)
 	if err := srv.Start(ctx); err != nil {
 		return fmt.Errorf("failed to start qbit server: %w", err)
 	}